requirements.txt
# REQUIREMENTS.TXT

tensorflow
tflite-runtime
opencv-python
flask
numpy
picamera2
RPi.GPIO
